New obj and map fix
Back-end only, no changelog.

Server files could not compile due to non-existent subtype of the
smartfridge in chemistry. This was replaced with a table.

New object: /obj/fixthisshit
It's a rather visible object you can replace missing types with when
looking at the map. If saved to the map, it's already named and
described to let players know it doesn't exist. It should look out of
place no matter where you put it. Use this if you don't know where the
missing type is located to make sensible replacements.