Character Setup unsets ready status.
This commit fixes a bug where declaring ready after round start causes
the character setup to lock and the button to become non-functional,
forcing the player to relog if they wish to change anything, by setting
the player's ready to 0 if when setup is accessed. This also saves the
player a click for last-minute changes.