Emergency Cloning Bugfix
Fixed a bug that made it impossible to clone from a corpse.

stat == 2 is the definition of death, as defined in setup.dm, and used
in all other code involving stat. Any change to this requires redefining
DEAD and confirming all code that checks for death complies with the
change.